KE-1052 Move `PykeRetry` to separate file
